Working on updating active graphs based on loaded mappings
